% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprints.R
\name{get_preprints}
\alias{get_preprints}
\alias{get_preprint}
\title{Fetch one or multiple preprints}
\usage{
get_preprints(
  search = NULL,
  limit = 10,
  communities = NULL,
  filters = NULL,
  tags = NULL,
  ...
)

get_preprint(pid, ...)
}
\arguments{
\item{search}{character search string}

\item{limit}{numeric limit results to. Default is 10}

\item{communities}{character string for name of community}

\item{filters}{other character string}

\item{tags}{character string for tags}

\item{...}{other arguments passed to httr::GET}

\item{pid}{doi or uuid of preprint}
}
\value{
list of preprints

list with a single preprint
}
\description{
Fetch one or multiple preprints
}
