% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviews.R
\name{get_reviews}
\alias{get_reviews}
\alias{get_review}
\alias{get_preprint_reviews}
\title{Fetch one or multiple reviews}
\usage{
get_reviews(limit = 10, full = FALSE, ...)

get_review(rid, full = FALSE, ...)

get_preprint_reviews(pid, limit = 10, full = FALSE, ...)
}
\arguments{
\item{limit}{numeric limit results to. Default is 10}

\item{full}{logical. Should full reviews or only rapid reviews be fetched. Default is FALSE}

\item{...}{other arguments passed to httr::GET}

\item{rid}{uuid of the review}

\item{pid}{doi or uuid of the preprint}
}
\value{
list of preprints

list with a single preprint

list with a single preprint
}
\description{
Fetch one or multiple reviews
}
